colaboraciones

+++ app/Http/Controllers/traitCRUD.php
@@ -45,9 +45,9 @@ trait traitCRUD{
     }
     
     protected function borrarFichero($fichero){
-        if (!isset($fichero) && strlen($fichero)<3) return null;
+        if (!isset($fichero) || strlen($fichero)<3) return null;
         if (file_exists($fichero)) unlink($fichero);
-        if (file_exists(storage_path('/app/' . $fichero))) unlink(storage_path('/app/' . $fichero));
+        if (file_exists(storage_path('app/' . $fichero))) unlink(storage_path('app/' . $fichero));
     }
     
     /*
